# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Image to Website API",
    "description": "Production backend for turning a single product photograph into a complete, working online store with deterministic Jinja rendering, AI vision intelligence, and server-managed commerce.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/templates/{slug}": {
      "get": {
        "tags": [
          "Design Packs"
        ],
        "summary": "Template Preview",
        "description": "One design, rendered with a demo seller of the kind it was built for.",
        "operationId": "template_preview_templates__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/triage": {
      "post": {
        "tags": [
          "Generation"
        ],
        "summary": "Triage",
        "description": "FR-2 / FR-3 / FR-5 \u2014 verdict + coaching + Product Spec, BEFORE generating.\n\nThe local quality pass is pure PIL and returns in well under 2s; the\ndetection call is what the rest of the wait is.",
        "operationId": "triage_triage_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_triage_triage_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/generate-from-photo": {
      "post": {
        "tags": [
          "Generation"
        ],
        "summary": "Generate From Photo",
        "description": "Runs the photo-first pipeline on a photo already cleared by /triage.",
        "operationId": "generate_from_photo_generate_from_photo_post",
        "requestBody": {
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/Body_generate_from_photo_generate_from_photo_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/job-status/{job_id}": {
      "get": {
        "tags": [
          "Generation"
        ],
        "summary": "Job Status",
        "operationId": "job_status_job_status__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/{website_id}/publish": {
      "post": {
        "tags": [
          "Publishing"
        ],
        "summary": "Publish Site",
        "operationId": "publish_site_api__website_id__publish_post",
        "parameters": [
          {
            "name": "website_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Website Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Publishing"
        ],
        "summary": "Publish Status",
        "operationId": "publish_status_api__website_id__publish_get",
        "parameters": [
          {
            "name": "website_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Website Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Publishing"
        ],
        "summary": "Unpublish Site",
        "operationId": "unpublish_site_api__website_id__publish_delete",
        "parameters": [
          {
            "name": "website_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Website Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/s/{slug}/": {
      "get": {
        "tags": [
          "Publishing"
        ],
        "summary": "Published Home",
        "operationId": "published_home_s__slug___get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/s/{slug}": {
      "get": {
        "tags": [
          "Publishing"
        ],
        "summary": "Published Home",
        "operationId": "published_home_s__slug__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/s/{slug}/{filename}": {
      "get": {
        "tags": [
          "Publishing"
        ],
        "summary": "Published Page",
        "description": "Serve a published page. This is the buyer-facing URL \u2014 the one that goes on\na WhatsApp status or a card \u2014 so it is deliberately short and readable.",
        "operationId": "published_page_s__slug___filename__get",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Slug"
            }
          },
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_generate_from_photo_generate_from_photo_post": {
        "properties": {
          "photo_token": {
            "type": "string",
            "title": "Photo Token"
          },
          "extra_tokens": {
            "type": "string",
            "title": "Extra Tokens",
            "default": ""
          },
          "brand_name": {
            "type": "string",
            "title": "Brand Name",
            "default": ""
          },
          "price": {
            "type": "string",
            "title": "Price",
            "default": ""
          },
          "seller_facts": {
            "type": "string",
            "title": "Seller Facts",
            "default": ""
          },
          "density": {
            "type": "string",
            "title": "Density",
            "default": "generous"
          },
          "spin": {
            "type": "integer",
            "title": "Spin",
            "default": 0
          },
          "override_triage": {
            "type": "boolean",
            "title": "Override Triage",
            "default": false
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "en"
          },
          "category": {
            "type": "string",
            "title": "Category",
            "default": ""
          },
          "sub_type": {
            "type": "string",
            "title": "Sub Type",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "photo_token"
        ],
        "title": "Body_generate_from_photo_generate_from_photo_post"
      },
...
      "Body_triage_triage_post": {
        "properties": {
          "photo": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Photo"
          },
          "photos": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Photos",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "photo"
        ],
        "title": "Body_triage_triage_post"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}